services/post/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-CORS==4.0.0
Flask-JWT-Extended==4.5.3
PyMySQL==1.1.0
cryptography==41.0.7
Werkzeug==2.3.7
python-dotenv==1.0.0
